+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 ## [0.01]
 
 - Initial release
+  
+## [0.02]
+
+- The latest lang server version in the config is downloaded. 
+
+## [0.03]
+
+- decoupled the lang server update with the extension
+- Added telemetry, disable by disabling the telemetry from vscode settings.